fix MSVC/MinGW compiler warnings
and also fix a compiler warning fixed earlier in a different
(and better) way, see #4963

+++ mscore/palette/palettemodel.cpp
@@ -193,7 +193,7 @@ QModelIndex PaletteTreeModel::parent(const QModelIndex& modelIndex) const
 int PaletteTreeModel::rowCount(const QModelIndex& parent) const
       {
       if (!parent.isValid())
-            return palettes().size();
+            return int(palettes().size());
 
       void* iptr = parent.internalPointer();
 
@@ -755,8 +755,8 @@ void PaletteTreeModel::updateCellsState(const Selection& sel, bool deactivateAll
       if (!sel.isSingle() || !cr)
             deactivateAll = true;
 
-      const int npalettes = palettes().size();
-      for (int row = 0; row < npalettes; ++row) {
+      const size_t npalettes = palettes().size();
+      for (size_t row = 0; row < npalettes; ++row) {
             PalettePanel* palette = palettes()[row].get();
             // TODO: should this be turned on for all palettes?
             if (palette->type() != PalettePanel::Type::Beam)
@@ -772,7 +772,7 @@ void PaletteTreeModel::updateCellsState(const Selection& sel, bool deactivateAll
                         }
                   }
 
-            const QModelIndex parent = index(row, 0, QModelIndex());
+            const QModelIndex parent = index(int(row), 0, QModelIndex());
             const QModelIndex first = index(0, 0, parent);
             const QModelIndex last = index(palette->ncells() - 1, 0, parent);
             emit dataChanged(first, last, { CellActiveRole });
